bulk-telegram-sticker/requirements.txt
telethon>=1.28.5
python-dotenv>=0.19.0
Pillow>=10.0.0 